BDD/requirements.txt
Django==5.1.2
django-jazzmin==3.0.1
django-multiselectfield==0.1.13
djangorestframework==3.12.4
django-filter==21.1